# docs/swagger.yaml
components:
  schemas:
    Character:
      properties:
        class:
          type: string
        current:
          description: Current always true if present
          type: boolean
        deleted:
          description: Deleted always true if present
          type: boolean
        equipment:
          items:
            $ref: '#/components/schemas/Item'
          type: array
          uniqueItems: false
        experience:
          type: integer
        expired:
          description: Expired always true if present
          type: boolean
        guardian:
          description: Guardian PoE1 only
          items:
            $ref: '#/components/schemas/Item'
          type: array
          uniqueItems: false
        id:
          description: Id a unique 64 digit hexadecimal string
          type: string
        inventory:
          description: Inventory PoE1 only
          items:
            $ref: '#/components/schemas/Item'
          type: array
          uniqueItems: false
        jewels:
          items:
            $ref: '#/components/schemas/Item'
          type: array
          uniqueItems: false
        league:
          type: string
        level:
          type: integer
        metadata:
          $ref: '#/components/schemas/CharacterMetadata'
        name:
          type: string
        passives:
          $ref: '#/components/schemas/CharacterPassives'
        realm:
          $ref: '#/components/schemas/Realm'
        rucksack:
          description: Rucksack PoE1 only
          items:
            $ref: '#/components/schemas/Item'
          type: array
          uniqueItems: false
        ruthless:
          description: Ruthless PoE1 only; always true if present
          type: boolean
        skills:
          description: Skills PoE2 only
          items:
            $ref: '#/components/schemas/Item'
          type: array
          uniqueItems: false
      required:
      - class
      - experience
      - id
      - level
      - name
      - realm
      type: object
    CharacterMetadata:
      properties:
        version:
          description: Version game version for the character's realm
          type: string
      type: object
    CharacterPassives:
      properties:
        alternate_ascendancy:
          description: AlternateAscendancy PoE1 only; Bloodline class name
          type: string
        bandit_choice:
          $ref: '#/components/schemas/CharacterPassivesBanditChoice'
        hashes:
          items:
            type: integer
          type: array
          uniqueItems: false
        hashes_ex:
          description: HashesEx PoE1 only
          items:
            type: integer
          type: array
          uniqueItems: false
        jewel_data:
          additionalProperties:
            $ref: '#/components/schemas/ItemJewelData'
          description: JewelData the key is the string value of the x property of
            an item from the jewels array in this request
          type: object
        mastery_effects:
          additionalProperties:
            type: integer
          description: MasteryEffects PoE1 only; the key is the string value of the
            mastery node skill hash and the value is the selected effect hash
          type: object
        pantheon_major:
          $ref: '#/components/schemas/CharacterPassivesPantheonMajor'
        pantheon_minor:
          $ref: '#/components/schemas/CharacterPassivesPantheonMinor'
        quest_stats:
          description: QuestStats PoE2 only; passives granted via quests
          items:
            type: string
          type: array
          uniqueItems: false
        skill_overrides:
          additionalProperties:
            $ref: '#/components/schemas/PassiveNode'
          description: SkillOverrides the key is the string value of the node identifier
            being replaced
          type: object
        specialisations:
          additionalProperties:
            items:
              type: integer
            type: array
          description: Specialisations PoE2 only; the keys are set1, set2, and set3
          type: object
      required:
      - hashes
      - hashes_ex
      - jewel_data
      - mastery_effects
      - skill_overrides
      - specialisations
      type: object
...
    Item:
      properties:
        abyssJewel:
          description: AbyssJewel always true if present
          type: boolean
        additionalProperties:
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        artFilename:
          type: string
        baseType:
          type: string
        bondedMods:
          description: BondedMods PoE2 only
          items:
            type: string
          type: array
          uniqueItems: false
        builtInSupport:
          description: BuiltInSupport PoE1 only; Supported by level 1 x
          type: string
        cisRaceReward:
          description: CisRaceReward always true if present
          type: boolean
        colour:
          $ref: '#/components/schemas/ItemColour'
        corrupted:
          description: Corrupted always true if present
          type: boolean
        cosmeticMods:
          items:
            type: string
          type: array
          uniqueItems: false
        crucible:
          $ref: '#/components/schemas/ItemCrucible'
        crucibleMods:
          description: CrucibleMods only allocated mods are included
          items:
            type: string
          type: array
          uniqueItems: false
        delve:
          description: Delve always true if present
          type: boolean
        descrText:
          type: string
        desecrated:
          description: Desecrated PoE2 only; always true if present
          type: boolean
        doubleCorrupted:
          description: DoubleCorrupted PoE2 only; always true if present
          type: boolean
        duplicated:
          description: Duplicated always true if present
          type: boolean
        elder:
          description: Elder always true if present
          type: boolean
        enchantMods:
          items:
            type: string
          type: array
          uniqueItems: false
        enshrouded:
          $ref: '#/components/schemas/ItemEnshrouded'
        explicitMods:
          items:
            $ref: '#/components/schemas/ItemMod'
          type: array
          uniqueItems: false
        extended:
          $ref: '#/components/schemas/ItemExtended'
        flavourText:
          items:
            type: string
          type: array
          uniqueItems: false
        flavourTextNote:
          description: FlavourTextNote user-generated text
          type: string
        foilVariation:
          type: integer
        foreseeing:
          description: Foreseeing always true if present
          type: boolean
        forum_note:
          description: ForumNote user-generated text
          type: string
        fractured:
          description: Fractured always true if present
          type: boolean
        frameType:
          description: FrameType deprecated; use frameTypeId
          type: integer
        frameTypeId:
          type: string
        gemBackground:
          description: GemBackground PoE2 only
          type: string
        gemSkill:
          description: GemSkill PoE2 only
          type: string
        gemSockets:
          description: GemSockets PoE2 only; string is always W
          items:
            type: string
          type: array
          uniqueItems: false
        gemTabs:
          description: GemTabs PoE2 only
          items:
            $ref: '#/components/schemas/GemTab'
          type: array
          uniqueItems: false
        grantedSkills:
          description: GrantedSkills PoE2 only
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        h:
          type: integer
        hybrid:
          $ref: '#/components/schemas/ItemHybrid'
        icon:
          type: string
        iconStackLevel:
          description: IconStackLevel describes the level of a stack of items
          type: string
        iconTierText:
          description: IconTierText usually roman numerals
          type: string
        id:
          description: Id a unique 64 digit hexadecimal string
          type: string
        identified:
          type: boolean
        ilvl:
          type: integer
        implicitMods:
          items:
            $ref: '#/components/schemas/ItemMod'
          type: array
          uniqueItems: false
        incubatedItem:
          $ref: '#/components/schemas/ItemIncubatedItem'
        influences:
          additionalProperties: {}
          type: object
        inventoryId:
          type: string
        isRelic:
          description: IsRelic always true if present
          type: boolean
        itemLevel:
          description: ItemLevel used for items that always display their item level
          type: integer
        league:
          type: string
        lockedToAccount:
          description: LockedToAccount always true if present
          type: boolean
        lockedToCharacter:
          description: LockedToCharacter always true if present
          type: boolean
        logbookMods:
          items:
            $ref: '#/components/schemas/ItemLogbookMod'
          type: array
          uniqueItems: false
        maxStackSize:
          type: integer
        memoryItem:
          description: MemoryItem always true if present
          type: boolean
        mercenarySkills:
          items:
            $ref: '#/components/schemas/ItemMercenarySkill'
          type: array
          uniqueItems: false
        monsterLevel:
          description: MonsterLevel PoE1 only; used for items that always display
            their monster level
          type: integer
        mutated:
          description: 'Mutated PoE1: true on Foulborn Uniques, PoE2: true on all
            Vaal Uniques'
          type: boolean
        name:
          type: string
        nextLevelRequirements:
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        notableProperties:
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        note:
          description: Note user-generated text
          type: string
        properties:
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        prophecyText:
          type: string
        rarity:
          $ref: '#/components/schemas/ItemRarity'
        realm:
          $ref: '#/components/schemas/Realm'
        replica:
          description: Replica always true if present
          type: boolean
        requirements:
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        rewards:
          items:
            $ref: '#/components/schemas/ItemReward'
          type: array
          uniqueItems: false
        runeMods:
          description: RuneMods PoE2 only
          items:
            type: string
          type: array
          uniqueItems: false
        ruthless:
          description: Ruthless always true if present
          type: boolean
        sanctified:
          description: Sanctified PoE2 only; always true if present
          type: boolean
        scourgeMods:
          items:
            type: string
          type: array
          uniqueItems: false
        scourged:
          $ref: '#/components/schemas/ItemScourged'
        seaRaceReward:
          description: SeaRaceReward always true if present
          type: boolean
        searing:
          description: Searing always true if present
          type: boolean
        secDescrText:
          type: string
        shaper:
          description: Shaper always true if present
          type: boolean
        socket:
          type: integer
        socketedIcon:
          description: SocketedIcon PoE2 only; an image URL to use when this item
            is in the socket of another item
          type: string
        socketedItems:
          items:
            $ref: '#/components/schemas/Item'
          type: array
          uniqueItems: false
        sockets:
          items:
            $ref: '#/components/schemas/ItemSocket'
          type: array
          uniqueItems: false
        split:
          description: Split always true if present
          type: boolean
        stackSize:
          type: integer
        stackSizeText:
          type: string
        support:
          description: Support always true if present
          type: boolean
        supportGemRequirements:
          description: SupportGemRequirements PoE2 only
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        synthesised:
          description: Synthesised always true if present
          type: boolean
        tamedBeastProperties:
          description: TamedBeastProperties PoE2 only
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        tangled:
          description: Tangled always true if present
          type: boolean
        thRaceReward:
          description: ThRaceReward always true if present
          type: boolean
        typeLine:
          type: string
        ultimatumMods:
          items:
            $ref: '#/components/schemas/ItemUltimatumMod'
          type: array
          uniqueItems: false
        unidentifiedTier:
          description: UnidentifiedTier PoE2 only
          type: integer
        unmodifiable:
          description: Unmodifiable always true if present
          type: boolean
        unmodifiableExceptChaos:
          description: UnmodifiableExceptChaos always true if present
          type: boolean
        utilityMods:
          items:
            type: string
          type: array
          uniqueItems: false
        veiled:
          description: Veiled always true if present
          type: boolean
        veiledMods:
          description: VeiledMods random video identifier
          items:
            type: string
          type: array
          uniqueItems: false
        verified:
          type: boolean
        vestigial:
          description: Vestigial PoE1 only; always true if present
          type: boolean
        w:
          type: integer
        weaponRequirements:
          description: WeaponRequirements PoE2 only
          items:
            $ref: '#/components/schemas/ItemProperty'
          type: array
          uniqueItems: false
        x:
          type: integer
        "y":
          type: integer
      required:
      - baseType
      - frameTypeId
      - h
      - icon
      - identified
      - ilvl
      - name
      - typeLine
      - verified
      - w
      type: object
...
    BplCharacter:
      properties:
        ascendancy:
          type: string
        ascendancy_points:
          type: integer
        atlas_node_count:
          type: integer
        event_id:
          type: integer
        id:
          type: string
        level:
          type: integer
        main_skill:
          type: string
        name:
          type: string
        pantheon:
          type: boolean
        user_id:
          type: integer
        void_stones:
          type: integer
      required:
      - ascendancy
      - ascendancy_points
      - atlas_node_count
      - event_id
      - id
      - level
      - main_skill
      - name
      - pantheon
      - void_stones
      type: object
...
    CharacterStat:
      properties:
        armour:
          type: integer
        attack_block:
          type: integer
        dps:
          type: integer
        ehp:
          type: integer
        ele_max_hit:
          type: integer
        es:
          type: integer
        evasion:
          type: integer
        high_level_flasks:
          type: integer
        hp:
          type: integer
        item_indexes:
          items:
            type: integer
          type: array
          uniqueItems: false
        lowest_ele_res:
          type: integer
        main_skill:
          type: string
        mana:
          type: integer
        movement_speed:
          type: integer
        phys_max_hit:
          type: integer
        spell_block:
          type: integer
        timestamp:
          type: integer
        xp:
          type: integer
      required:
      - armour
      - attack_block
      - dps
      - ehp
      - ele_max_hit
      - es
      - evasion
      - high_level_flasks
      - hp
      - item_indexes
      - lowest_ele_res
      - main_skill
      - mana
      - movement_speed
      - phys_max_hit
      - spell_block
      - timestamp
      - xp
      type: object
...
    PoB:
      properties:
        armour:
          type: integer
        ascendancy:
          type: string
        dps:
          type: integer
        ehp:
          type: integer
        ele_max_hit:
          type: integer
        es:
          type: integer
        evasion:
          type: integer
        export_string:
          type: string
        hp:
          type: integer
        id:
          type: integer
        level:
          type: integer
        main_skill:
          type: string
        mana:
          type: integer
        movement_speed:
          type: integer
        phys_max_hit:
          type: integer
        timestamp:
          format: date-time
          type: string
        xp:
          type: integer
      required:
      - armour
      - ascendancy
      - dps
      - ehp
      - ele_max_hit
      - es
      - evasion
      - export_string
      - hp
      - id
      - level
      - main_skill
      - mana
      - movement_speed
      - phys_max_hit
      - timestamp
      - xp
      type: object
...
    User:
      properties:
        account_name:
          type: string
        discord_id:
          type: string
        discord_name:
          type: string
        display_name:
          type: string
        id:
          type: integer
        permissions:
          items:
            $ref: '#/components/schemas/Permission'
          type: array
          uniqueItems: false
        token_expiry_timestamp:
          format: date-time
          type: string
        twitch_id:
          type: string
        twitch_name:
          type: string
      required:
      - display_name
      - id
      - permissions
      type: object
...
    Permission:
      enum:
      - admin
      - manager
      - objective_designer
      - submission_judge
      - achievement_assigner
      type: string
      x-enum-varnames:
      - PermissionAdmin
      - PermissionManager
      - PermissionObjectiveDesigner
      - PermissionSubmissionJudge
      - PermissionAchievementAssigner
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: Liberatorist@gmail.com
    name: Liberator
  description: This is the backend API for the BPL project.
  title: BPL Backend API
  version: "2.0"
openapi: 3.1.0
paths:
...
  /events/{event_id}/characters:
    get:
      description: Get all characters for an event
      operationId: GetCharactersForEvent
      parameters:
      - description: Event ID
        in: path
        name: event_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BplCharacter'
                type: array
          description: OK
      tags:
      - characters
...
  /users/{user_id}:
    get:
      description: Fetches a user by ID
      operationId: GetUserById
      parameters:
      - description: User Id
        in: path
        name: user_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
      tags:
      - user
    patch:
      description: Changes the permissions of a user
      operationId: ChangePermissions
      parameters:
      - description: User Id
        in: path
        name: user_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - items:
                  $ref: '#/components/schemas/Permission'
                title: permissions
                type: array
        description: Permissions
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
      security:
      - BearerAuth: []
      tags:
      - user
  /users/{user_id}/characters:
    get:
      description: Fetches all event characters for a user
      operationId: GetUserCharacters
      parameters:
      - description: User Id
        in: path
        name: user_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BplCharacter'
                type: array
          description: OK
      tags:
      - characters
  /users/{user_id}/characters/{character_id}:
    get:
      description: Get all character data for an event for a user
      operationId: GetCharacterHistory
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        schema:
          type: integer
      - description: Character ID
        in: path
        name: character_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CharacterStat'
                type: array
          description: OK
      tags:
      - characters
    patch:
      description: Update character details
      operationId: UpdateCharacter
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        schema:
          type: integer
      - description: Character ID
        in: path
        name: character_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
          description: OK
      tags:
      - characters
  /users/{user_id}/characters/{character_id}/pobs:
    get:
      description: Get all PoB exports for a character
      operationId: GetPoBs
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        schema:
          type: integer
      - description: Character ID
        in: path
        name: character_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PoB'
                type: array
          description: OK
      tags:
      - characters
  /users/{user_id}/characters/{character_id}/pobs/{pob_id}:
    delete:
      description: Delete a PoB export for a character
      operationId: DeletePoBExport
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        schema:
          type: integer
      - description: Character ID
        in: path
        name: character_id
        required: true
        schema:
          type: string
      - description: PoB Export ID
        in: path
        name: pob_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
      tags:
      - characters
...